Add tests and status for incubation process to prevent early incubation

+++ traits/src/lib.rs
@@ -42,5 +42,5 @@ pub mod primitives {
 	pub type ZIndex = u32;
 	pub type SerialId = u32;
 	pub type PreorderId = u32;
-	pub type EraId = u128;
+	pub type EraId = u64;
 }
